feat: Only close the navigation when resizing horizontally (#1344)
* feat: Only hide the navigation when resizing horizontally

* refactor: Move throttle into utils folder
