+++ static/js/navigation/main.js
@@ -1,3 +1,5 @@
+import throttle from "../utils/throttle";
+
 import {
   navigation,
   secondaryNavigation,
@@ -219,25 +221,18 @@ function closeAllNavigationItems({ exception } = {}) {
   }
 }
 
-/**
- * Throttle util (for window resize event)
- * @param {Function} fn
- * @param {Int} delay
- */
-var throttle = function (fn, delay) {
-  var timer = null;
-  return function () {
-    var context = this,
-      args = arguments;
-    clearTimeout(timer);
-    timer = setTimeout(function () {
-      fn.apply(context, args);
-    }, delay);
-  };
-};
-
-// hide navigation when screen is resized
-window.addEventListener("resize", throttle(closeAllNavigationItems, 10));
+// Hide navigation when screen is horizontally resized
+let previousWidth = window.innerWidth;
+window.addEventListener(
+  "resize",
+  throttle(function () {
+    const currentWidth = window.innerWidth;
+    if (currentWidth !== previousWidth) {
+      closeAllNavigationItems();
+      previousWidth = currentWidth;
+    }
+  }, 10)
+);
 
 // Update careers dropdown with latest avaiable roles
 populateCareersRoles();
